[level 2] Title: 멀쩡한 사각형, Time: 0.01 ms, Memory: 3.64 MB -BaekjoonHub

+++ 프로그래머스/2/62048. 멀쩡한 사각형/멀쩡한 사각형.cpp
@@ -8,11 +8,9 @@ int gcd(int a, int b){
 }
 
 long long solution(int w,int h) {
-    //if(w == h) return w * h - w;
-    
     int cnt = gcd(w, h);
     int x = w / cnt;
     int y = h / cnt;
     
-    return ((ll)w * (ll)h) - ((ll)x + (ll)y - 1) * (ll)cnt;
+    return ((ll)w * (ll)h) - (x + y - 1) * cnt;
 }
